allow for setting ids on toasts so they can be replaced rather than added onto for repeat actions
